chore: some cleanup of files
