--- Attributes/AliasesAttribute.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace DeepDungeonDex.Attributes
-{
-    [AttributeUsage(AttributeTargets.Method)]
-    public class AliasesAttribute : Attribute
-    {
-        public string[] Aliases { get; }
-
-        public AliasesAttribute(params string[] aliases)
-        {
-            Aliases = aliases;
-        }
-    }
-}
