--- Attributes/CommandAttribute.cs
@@ -1,15 +1,0 @@
-using System;
-
-namespace DeepDungeonDex.Attributes
-{
-    [AttributeUsage(AttributeTargets.Method)]
-    public class CommandAttribute : Attribute
-    {
-        public string Command { get; }
-
-        public CommandAttribute(string command)
-        {
-            Command = command;
-        }
-    }
-}
