--- Attributes/DoNotShowInHelpAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace DeepDungeonDex
-{
-    [AttributeUsage(AttributeTargets.Method)]
-    public class DoNotShowInHelpAttribute : Attribute
-    {
-    }
-}
