--- Attributes/HelpMessageAttribute.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace DeepDungeonDex.Attributes
-{
-    [AttributeUsage(AttributeTargets.Method)]
-    public class HelpMessageAttribute : Attribute
-    {
-        public string HelpMessage { get; }
-
-        public HelpMessageAttribute(string helpMessage)
-        {
-            HelpMessage = helpMessage;
-        }
-    }
-}
